modif des classes ventes

+++ Programmation/E-Market/Commerce/Produit/Vente/Vente.cpp
@@ -2,7 +2,7 @@
 
 Vente::Vente(){}
 
-int Vente::getPrix(){}
+float Vente::getPrix(){}
 
 void Vente::setPrix(float newPrix){}
